+++ packages/core/src/shapes/Line.ts
@@ -1,8 +1,7 @@
 import { constOf } from "engine/Autodiff";
-import { bboxFromLinelike } from "engine/BBox";
 import { VarAD } from "types/ad";
+import { INamed, IStroke, IArrow, IShape } from "types/shapes";
 import { IVectorV } from "types/value";
-import { IArrow, INamed, IShape, IStroke, ShapeDef } from "./Shapes";
 import {
   BoolV,
   Canvas,
@@ -38,12 +37,4 @@ export const makeLine = (canvas: Canvas, properties: Partial<ILine>): Line => ({
   ...sampleLine(canvas),
   ...properties,
   shapeType: "Line",
-});
-
-export const Line = ShapeDef({
-  sampler: sampleLine,
-  constr: makeLine,
-
-  bbox: bboxFromLinelike,
-  isLinelike: true,
 });
